rename ui components to lowercase for consistency
